test unitarios

+++ user_microservice/src/main/java/com/pragma/user_microservice/domain/usecase/UserUseCase.java
@@ -6,9 +6,12 @@ import com.pragma.user_microservice.domain.exception.DomainException;
 import com.pragma.user_microservice.domain.model.CommonResponseModel;
 import com.pragma.user_microservice.domain.model.UserModel;
 import com.pragma.user_microservice.domain.spi.IUserPersistencePort;
-import com.pragma.user_microservice.infrastructure.configuration.Constants;
-import com.pragma.user_microservice.infrastructure.exception.NoDataFoundException;
+import com.pragma.user_microservice.domain.exception.AgeNotAllowedException;
 
+import java.time.LocalDate;
+import java.time.Period;
+import java.time.ZoneId;
+import java.time.format.DateTimeFormatter;
 import java.util.List;
 
 public class UserUseCase implements IUserServicePort {
@@ -20,8 +23,25 @@ public class UserUseCase implements IUserServicePort {
     }
 
     @Override
-    public void saveUser(UserModel userModel) {
-        iUserPersistencePort.saveUser(userModel);
+    public UserModel saveUser(UserModel userModel) {
+
+        validateAge(userModel);
+
+        return iUserPersistencePort.saveUser(userModel);
+    }
+
+    public void validateAge(UserModel userModel) {
+        DateTimeFormatter fmt = DateTimeFormatter.ofPattern("yyyy-dd-MM");
+        LocalDate birthDate = LocalDate.parse(userModel.getBirthDate().toInstant()
+                .atZone(ZoneId.systemDefault())
+                .toLocalDate().toString(), fmt);
+        LocalDate dateNow = LocalDate.now();
+
+        Period period = Period.between(birthDate, dateNow);
+
+        if(period.getYears() < 18){
+            throw new AgeNotAllowedException();
+        }
     }
 
     @Override
@@ -45,13 +65,21 @@ public class UserUseCase implements IUserServicePort {
     }
 
     @Override
-    public void updateUser(UserModel userModel) {
-        iUserPersistencePort.updateUser(userModel);
+    public UserModel updateUser(UserModel userModel) {
+        if(getUser(userModel.getId()) == null)
+            throw new DomainException(new CommonResponseDto("404","User not found.", false).toString());
+
+        return iUserPersistencePort.updateUser(userModel);
     }
 
     @Override
-    public void deleteUser(Long userId) {
+    public CommonResponseModel deleteUser(Long userId) {
+        if(getUser(userId) == null)
+            throw new DomainException(new CommonResponseDto("404","User not found.", false).toString());
+
         iUserPersistencePort.deleteUser(userId);
+
+        return new CommonResponseModel("200", "OK.", true);
     }
 
 }
